requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
